end of v2 sad face

+++ get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: earnaud <earnaud@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/17 12:49:51 by earnaud           #+#    #+#             */
-/*   Updated: 2020/11/21 13:01:43 by earnaud          ###   ########.fr       */
+/*   Updated: 2020/11/22 10:43:38 by earnaud          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ char    *ft_strjoin(char const *s1, char const *s2);
 size_t  ft_strlcpy(char *dest, const char *src, size_t size);
 size_t	ft_strlen(const char *str);
 char	*ft_strdup(const char *s);
-size_t	ft_line(const char *s);
+long int	ft_line(const char *s);
 char	*ft_afterline(char *s);
 void	ft_putstr_fd(char *s, int fd);
